Shift downloadFile from mobx store to utils
