+++ apps/studio/components/to-be-cleaned/Storage/StorageExplorer/FileExplorerHeaderSelection.tsx
@@ -1,17 +1,19 @@
 import { Download, Move, Trash2, X } from 'lucide-react'
 
 import { PermissionAction } from '@supabase/shared-types/out/constants'
+import { useParams } from 'common'
 import { ButtonTooltip } from 'components/ui/ButtonTooltip'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
 import { useStorageStore } from 'localStores/storageExplorer/StorageExplorerStore'
 import { Button } from 'ui'
+import { downloadFile } from './StorageExplorer.utils'
 
 const FileExplorerHeaderSelection = () => {
+  const { ref: projectRef, bucketId } = useParams()
   const canUpdateFiles = useCheckPermissions(PermissionAction.STORAGE_WRITE, '*')
   const storageExplorerStore = useStorageStore()
   const {
     selectedItems,
-    downloadFile,
     downloadSelectedFiles,
     clearSelectedItems,
     setSelectedItemsToDelete,
@@ -35,7 +37,7 @@ const FileExplorerHeaderSelection = () => {
           type="primary"
           onClick={async () => {
             if (selectedItems.length === 1) {
-              await downloadFile(selectedItems[0])
+              await downloadFile({ projectRef, bucketId, file: selectedItems[0] })
             } else {
               await downloadSelectedFiles(selectedItems)
             }
